suhdood/requirements.txt
Django==1.10.5
django-friendship==1.5.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
PyJWT==1.4.2
